refactor tests, enable testing in ci
