+++ relayer/middleware/missedVaas/index.ts
@@ -1,0 +1,2 @@
+export { missedVaas } from "./carefulMissedVaas.middleware";
+export type { MissedVaaOpts } from "./carefulMissedVaas.middleware";
